requirements.txt
pip~=20.2.4
setuptools~=50.3.2
requests~=2.21.0
retranslator~=0.1.2
websocket-client~=0.57.0
regex~=2020.10.23
beautifulsoup4~=4.9.3
aiohttp~=3.6.3
idna~=2.6
pycparser~=2.20
cryptography~=3.1.1
urllib3~=1.24.1
chardet~=3.0.4
certifi~=2018.8.24
six~=1.12.0
multidict~=4.7.6
future~=0.18.2
tornado~=6.0.4
decorator~=4.4.2
pycurl~=7.43.0.2
Pillow~=7.0.0
saya~=0.3.2